Proper indentation formatting now.
This is compliant, as far as I'm aware? This basically handles all of the aspects of the T&Q I can think of for memorandum usage.

There might be some ambiguity with adding a signature block, but that can't possibly be more difficult than the rest of this. 10/10, would do again.